Repackage {io.bisq.monitor => bisq.monitor}
