+++ pom.xml
@@ -85,7 +85,7 @@
                     <transformers>
                         <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                             <manifestEntries>
-                                <Main-Class>io.bisq.monitor.MonitorMain</Main-Class>
+                                <Main-Class>bisq.monitor.MonitorMain</Main-Class>
                                 <!-- the specified bouncy castle jar classes  -->
                                 <Class-Path>lib/bcpg-jdk15on.jar lib/bcprov-jdk15on.jar</Class-Path>
                             </manifestEntries>
