+++ src/main/java/bisq/monitor/MonitorAppSetup.java
@@ -15,7 +15,7 @@
  * along with bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.bisq.monitor;
+package bisq.monitor;
 
 import bisq.common.app.Version;
 import bisq.common.crypto.KeyRing;
@@ -23,7 +23,7 @@ import bisq.common.proto.persistable.PersistedDataHost;
 import bisq.core.app.BisqEnvironment;
 import bisq.core.app.SetupUtils;
 import bisq.core.btc.wallet.WalletsSetup;
-import io.bisq.monitor.metrics.p2p.MonitorP2PService;
+import bisq.monitor.metrics.p2p.MonitorP2PService;
 import bisq.network.crypto.EncryptionService;
 import bisq.network.p2p.network.SetupListener;
 import bisq.network.p2p.peers.PeerManager;
