+++ src/main/java/bisq/monitor/MonitorOptionKeys.java
@@ -15,24 +15,11 @@
  * along with bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.bisq.monitor.metrics;
+package bisq.monitor;
 
-import lombok.Getter;
-import lombok.Setter;
+public class MonitorOptionKeys {
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
-@Getter
-public class Metrics {
-    List<Long> requestDurations = new ArrayList<>();
-    List<String> errorMessages = new ArrayList<>();
-    List<Map<String, Integer>> receivedObjectsList = new ArrayList<>();
-    @Setter
-    long lastDataRequestTs;
-    @Setter
-    long lastDataResponseTs;
-    @Setter
-    long numRequestAttempts;
+    public static final String SLACK_URL_SEED_CHANNEL = "slackUrlSeedChannel";
+    public static final String SLACK_BTC_SEED_CHANNEL = "slackUrlBtcChannel";
+    public static final String SLACK_PROVIDER_SEED_CHANNEL = "slackUrlProviderChannel";
 }
