+++ src/main/java/bisq/monitor/metrics/p2p/MonitorP2PModule.java
@@ -15,8 +15,9 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.bisq.monitor.metrics.p2p;
+package bisq.monitor.metrics.p2p;
 
+import bisq.monitor.metrics.MetricsModel;
 import com.google.inject.Singleton;
 import com.google.inject.name.Names;
 import bisq.common.app.AppModule;
@@ -32,7 +33,6 @@ import bisq.network.p2p.peers.getdata.RequestDataManager;
 import bisq.network.p2p.peers.keepalive.KeepAliveManager;
 import bisq.network.p2p.peers.peerexchange.PeerExchangeManager;
 import bisq.network.p2p.storage.P2PDataStorage;
-import io.bisq.monitor.metrics.MetricsModel;
 import org.springframework.core.env.Environment;
 
 import java.io.File;
